Support regex input mode
Prefix input text with '\v', '\m' or '\M' to turn on "regex mode".

This doesn't break search of "literal text" whose length is less than 3.
Because only text matching '^\\[vVmM].' are not literal any more, which
is identical to what "literal text" means in Vim search pattern starting
with '\V'.

Systemic adjustments are needed for utilizing "regex mode", which is
better to be served by the author.

Close #104.